billion-token-system/requirements.txt
python-dotenv==1.0.0
aiohttp==3.9.1
asyncio-throttle==1.0.2
redis==5.0.1
litellm==1.46.0
rich==13.7.0
prometheus-client==0.19.0
pydantic==2.5.3
tenacity==8.2.3
cryptography==41.0.7